Answer to these comments in the commit message
The parent directory attribute was added and all the work happens on filenames vs filepaths because it easier to handle situations when you want to access the file but the current directory (as in the directory that the python interpreter is in executing in; controllable with `os.chdir`) is not the same as where the script was started. To get the filepath, you can simply appent filename to parent directory.